fix(anyharness): satisfy repo-shape line budget
- split fork-operation model types into their own module (model.rs back under 600)
- move fork taxonomy test to a sibling file; compact remaining fork test
- bump max-lines allowlist for the oversized contract + runtime test files
