+++ anyharness/crates/anyharness-lib/src/domains/sessions/mod.rs
@@ -7,6 +7,7 @@ pub mod delegation;
 pub mod deletion;
 pub mod execution_summary;
 pub mod extensions;
+pub mod fork_operation;
 pub mod links;
 pub mod live_config;
 pub mod live_ports;
