+++ scripts/max_lines_allowlist.txt
@@ -8,7 +8,7 @@
 # when the file is under its threshold.
 
 anyharness/crates/anyharness-contract/src/v1/events.rs 1092 existing AnyHarness contract oversized file
-anyharness/crates/anyharness-contract/src/v1/sessions.rs 696 existing AnyHarness contract oversized file (+caller-selected idempotent create id; stale fingerprint reduced after legacy public subagent response removal)
+anyharness/crates/anyharness-contract/src/v1/sessions.rs 703 existing AnyHarness contract oversized file (+caller-selected idempotent create id; stale fingerprint reduced after legacy public subagent response removal; +Forks ADR rung 2 caller-reserved fork child_session_id)
 anyharness/crates/anyharness-lib/src/api/openapi.rs 683 existing AnyHarness oversized file (+hosting schema registration, +agent-auth state response, +gateway catalog endpoints) + enriched gateway-models schema fields (model-table §1) + workflow-runs + local materialization operations + workflow-run-workspaces paths/schemas + agent install progress schemas + worktree restore path/schemas + WorktreeBaseFetch + RuntimeDiskPressure schemas (subagent registrations extracted to their owned OpenAPI document) ; -10: R1 archiving deletes the three retire path registrations, the three worktree-retention path registrations and the three WorkspaceRetire* schema registrations ; +22: R4 archiving registers the two lifecycle paths and the fourteen lifecycle schemas ; -4: R5 deletes the purge-preflight/retry path registrations and the WorkspaceCleanupOperation/WorkspacePurgePreflightResponse schema registrations ; -30: workflows gen-2 PR1 deletes the gen-1 workflow-runs and workflow-run-workspaces path/schema registrations ; +2: PR5a merges the workflow-runs OpenAPI document into the composed doc ; +16: agent-auth canonical evidence model registers the thirteen AgentAuth* schemas
 anyharness/crates/anyharness-lib/src/api/router_tests.rs 1898 existing AnyHarness oversized test file (+hosting PR-status route tests, +catalog version tests, +managed default-placement route test, +model-snapshot status route shape/auth tests, +agent-auth/install poke contract tests, +agent-auth/install poke-suppression proofs against the automatic_poke_engine seam (F-036 cycle-2 review item 1); the catalog-push handler tests are gone with the heartbeat transport)
 anyharness/crates/anyharness-lib/src/domains/agent_operations/runtime/error.rs 618 agent-operations error mapping + outcome classifier (+18: #1849 review fixes classify WORKSPACE_ARCHIVED / SUBAGENT_FANOUT_LIMIT / WORKSPACE_SINGLE_SESSION as denied and extend the classifier test)
@@ -27,7 +27,7 @@ anyharness/crates/anyharness-lib/src/domains/goals/service_tests.rs 682 goal ser
 anyharness/crates/anyharness-lib/src/domains/reviews/service/service_tests.rs 1510 existing AnyHarness oversized test file
 anyharness/crates/anyharness-lib/src/domains/sessions/runtime/prompt.rs 609 session prompt dispatch ; +39: workflows gen-2 PR4 adds send_text_blocks_prompt_with_id (the workflow envelope's multi-block twin) sharing one send_payload_prompt_with_id seam with the text path
 anyharness/crates/anyharness-lib/src/domains/sessions/runtime/prompt_message_actor_tests.rs 601 scripted-ACP prompt/queue race suite over the real actor ; +2: workflows gen-2 PR4 teaches the scripted agent program a refusal stop-reason branch
-anyharness/crates/anyharness-lib/src/domains/sessions/runtime/tests.rs 1019 existing AnyHarness oversized test file (launch-env tests moved to their owning module) + A9 Scope C resume/fork revoked-credentials regression tests (AgentNotReady at the common live-start seam) + a review-fix pin (unsatisfiable selection reports RouteAuth, not AgentNotReady, on resume)
+anyharness/crates/anyharness-lib/src/domains/sessions/runtime/tests.rs 1213 existing AnyHarness oversized test file (launch-env tests moved to their owning module) + A9 Scope C resume/fork revoked-credentials regression tests (AgentNotReady at the common live-start seam) + a review-fix pin (unsatisfiable selection reports RouteAuth, not AgentNotReady, on resume) + Forks ADR rung 2 targeted-fork plumbing tests (item-less rejection, fail-closed no-child downgrade guard, idempotency conflict, unknown-outcome redispatch block)
 anyharness/crates/anyharness-lib/src/live/sessions/probe.rs 679 catalog probe with per-harness model-source fallbacks (+the three runtime-probe options: env removal, scratch workspace root, config-option-only switch suppression); shrank from 683 after the readiness resolve moved to the call boundary (grid plan PR 5c/7 — live never fetches)
 anyharness/crates/anyharness-lib/src/live/workflows/lifecycle_tests.rs 1070 workflows gen-2 engine lifecycle suite: scripted-ACP agents drive real runs through the manager/actor (happy path, gate approve, flip, redo, undo, boot fence) with zero mocks of our machinery
 anyharness/crates/anyharness-lib/src/live/terminals/manager_tests.rs 730 R3 workspace-stop terminal suite: +615 lines of stop_all_for_workspace coverage (kill census, setup-process reaping, wedged-terminal escalation) sharing one real-PTY manager harness per case
